requirements.txt
django-cors-headers==1.1.0
djangorestframework==3.1.3
requests==2.10.0
pytz==2016.6.1
django-bootstrap-form==3.2
django-braces==1.8.1
django-ckeditor==5.0.3
Pillow==3.3.1
markdown==2.6.6
django_filter==0.14.0
djangorestframework-jwt==1.8.0